scripts/requirements.txt
python-socketio
psychopg2
python-dotenv